sqm-scripts: Test 20230217
runners: Expose a 'cleanup' action for silent cleanup of an interface

If a network interface is removed before SQM is stopped on that interface,
removing the qdisc from that interface will result in errors in the log.
Whereas not running the 'stop' action will result in leftover state
files (and possibly an IFB interface).

While the right thing to do would be to stop SQM before removing the
interface, this is not always possible; for instance, pppd doesn't have a
'pre-down' hook. So to accommodate this use case, add a new 'cleanup'
action, which just unwinds the firewall rules, removes the IFB
interface (but silently in case it does not exist), and deletes the state
file.